# api-gateway/api/api-gateway-openapi.yaml
openapi: 3.0.0
info:
  title: DopMaterials
  description: Описание методов api-gateway
  version: "1.0.0"
servers:
 - url: "http://localhost:443"
paths:

  /photos/{filename_with_extension}:
    get:
      tags:
        - photos
      summary: Получить фотографию
      parameters:
        - name: filename_with_extension
          in: path
          required: true
          description: имя файла с расширением
          schema:
            type: string
      responses:
        '400':
          description: "Неправильные или неполные данные"
        '404':
          description: "Не найдены данные"
        '500':
          description: "Неизвестная ошибка сервера"
        '200':
          description: Фотография получена
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
    delete:
      tags:
        - photos
      summary: Получить фотографию
      parameters:
        - name: filename_with_extension
          in: path
          required: true
          description: имя файла с расширением
          schema:
            type: string
      responses:
        '400':
          description: "Неправильные или неполные данные"
        '404':
          description: "Не найдены данные"
        '500':
          description: "Неизвестная ошибка сервера"
        '204':
          description: Фотография удалена
  /photos/{filename}:
    post:
      tags:
        - photos
      summary: Добавить фотографию
      parameters:
        - name: filename
          in: path
          required: true
          description: имя файла
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                photo:
                  type: string
                  format: binary
      responses:
        '400':
          description: "Неправильные или неполные данные"
        '413':
          description: "Файл превышает допустимый размер"
        '500':
          description: "Неизвестная ошибка сервера"
        '201':
          description: Фотография добавлена
  /videos/{filename_with_extension}:
    get:
      tags:
        - videos
      summary: Получить видео
      parameters:
        - name: filename_with_extension
          in: path
          required: true
          description: имя файла с расширением
          schema:
            type: string
      responses:
        '400':
          description: "Неправильные или неполные данные"
        '404':
          description: "Не найдены данные"
        '500':
          description: "Неизвестная ошибка сервера"
        '200':
          description: Видео получено
          content:
            video/mp4:
              schema:
                type: string
                format: binary
    delete:
      tags:
        - videos
      summary: Получить видео
      parameters:
        - name: filename_with_extension
          in: path
          required: true
          description: имя файла с расширением
          schema:
            type: string
      responses:
        '400':
          description: "Неправильные или неполные данные"
        '404':
          description: "Не найдены данные"
        '500':
          description: "Неизвестная ошибка сервера"
        '204':
          description: Видео удалено
  /videos/{filename}:
    post:
      tags:
        - videos
      summary: Добавить видео
      parameters:
        - name: filename
          in: path
          required: true
          description: имя файла
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                video:
                  type: string
                  format: binary
      responses:
        '400':
          description: "Неправильные или неполные данные"
        '413':
          description: "Файл превышает допустимый размер"
        '500':
          description: "Неизвестная ошибка сервера"
        '201':
          description: Видео добавлено

  /materials/{id}:
    get:
      tags:
        - materials
      summary: Получить материал
      parameters:
        - name: id
          in: path
          required: true
          description: ID материала
          schema:
            type: integer
      responses:
        '400':
          description: "Неправильные или неполные данные"
        '404':
          description: "Не найдены данные"
        '500':
          description: "Неизвестная ошибка сервера"
        '200':
          description: Материал получен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Material'
    delete:
      tags:
        - materials
      summary: Удалить материал
      parameters:
        - name: id
          in: path
          required: true
          description: ID материала
          schema:
            type: integer
      responses:
        '400':
          description: "Неправильные или неполные данные"
        '404':
          description: "Не найдены данные"
        '500':
          description: "Неизвестная ошибка сервера"
        '204':
          description: Материал удален
  /materials/{id}/restore:
    patch:
      tags:
        - materials
      summary: Восстановить материал
      parameters:
        - name: id
          in: path
          required: true
          description: ID материала
          schema:
            type: integer
      responses:
        '400':
          description: "Неправильные или неполные данные"
        '404':
          description: "Не найдены данные"
        '500':
          description: "Неизвестная ошибка сервера"
        '204':
          description: Материал восстановлен
  /materials:
    get:
      tags:
        - materials
      summary: Получить список материалов
      parameters:
        - name: category_id
          in: query
          required: false
          description: ID категории материала
          schema:
            type: integer
        - name: find_part
          in: query
          required: false
          description: Часть названия для поиска
          schema:
            type: string
        - name: offset
          in: query
          required: false
          description: Смещение для пагинации
          schema:
            type: integer
      responses:
        '400':
          description: "Неправильные или неполные данные"
        '500':
          description: "Неизвестная ошибка сервера"
        '200':
          description: Список материалов получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Material'
    post:
      tags:
        - materials
      summary: Добавить материал
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaterialCreate'
      responses:
        '400':
          description: "Неправильные или неполные данные"
        '500':
          description: "Неизвестная ошибка сервера"
        '201':
          description: Материал добавлен
    put:
      tags:
        - materials
      summary: Обновить материал
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaterialUpdate'
      responses:
        '400':
          description: "Неправильные или неполные данные"
        '404':
          description: "Не найдены данные"
        '500':
          description: "Неизвестная ошибка сервера"
        '204':
          description: Материал обновлен

  /categories/{id}:
    get:
      tags:
        - categories
      summary: Получить категорию
      parameters:
        - name: id
          in: path
          required: true
          description: ID категории
          schema:
            type: integer
      responses:
        '400':
          description: "Неправильные или неполные данные"
        '404':
          description: "Не найдены данные"
        '500':
          description: "Неизвестная ошибка сервера"
        '200':
          description: Категория получена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Material'
    delete:
      tags:
        - categories
      summary: Удалить категорию
      parameters:
        - name: id
          in: path
          required: true
          description: ID категории
          schema:
            type: integer
      responses:
        '400':
          description: "Неправильные или неполные данные"
        '404':
          description: "Не найдены данные"
        '500':
          description: "Неизвестная ошибка сервера"
        '204':
          description: Категория удалена
  /categories/{id}/subsidiaries:
    get:
      tags:
        - categories
      summary: Получить дочерние категории
      parameters:
        - name: id
          in: path
          required: true
          description: ID категории
          schema:
            type: integer
      responses:
        '400':
          description: "Неправильные или неполные данные"
        '404':
          description: "Не найдены данные"
        '500':
          description: "Неизвестная ошибка сервера"
        '200':
          description: Дочерние категории получена
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
  /categories:
    get:
      tags:
        - categories
      summary: Получить список категорий
      responses:
        '400':
          description: "Неправильные или неполные данные"
        '500':
          description: "Неизвестная ошибка сервера"
        '200':
          description: Список категорий получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
    post:
      tags:
        - categories
      summary: Добавить категорию
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreate'
      responses:
        '400':
          description: "Неправильные или неполные данные"
        '500':
          description: "Неизвестная ошибка сервера"
        '201':
          description: Категория добавлена
    put:
      tags:
        - categories
      summary: Обновить категорию
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        '400':
          description: "Неправильные или неполные данные"
        '404':
          description: "Не найдены данные"
        '500':
          description: "Неизвестная ошибка сервера"
        '204':
          description: Категория обновлена


components:
  schemas:

    Material:
      type: object
      properties:
        id:
          type: integer
          example: 1
        category_id:
          type: integer
          example: 1
        name:
          type: string
          example: 'материал'
        description:
          type: string
          example: 'удивительный'
        preview_meta:
          type: string
          example: '1.png'
        video_meta:
          type: string
          example: '1'
        deleted:
          type: boolean
          example: 'false'
        created_at:
          type: string
          format: date-time
          example: 2024-06-26T07:17:45.583005Z
        updated_at:
          type: string
          format: date-time
          example: 2024-06-26T07:17:45.583005Z
    MaterialCreate:
      type: object
      properties:
        category_id:
          type: integer
          example: 1
        name:
          type: string
          example: 'материал'
        description:
          type: string
          example: 'удивительный'
        preview_meta:
          type: string
          example: '1.png'
        video_meta:
          type: string
          example: '1'
    MaterialUpdate:
      type: object
      properties:
        id:
          type: integer
          example: 1
        category_id:
          type: integer
          example: 1
        name:
          type: string
          example: 'материал'
        description:
          type: string
          example: 'удивительный'
        preview_meta:
          type: string
          example: '1.png'
        video_meta:
          type: string
          example: '1'

    Category:
      type: object
      properties:
        id:
          type: integer
          example: 1
        category_id:
          type: integer
          example: 1
        name:
          type: string
          example: 'категория'
    CategoryCreate:
      type: object
      properties:
        category_id:
          type: integer
          example: 1
        name:
          type: string
          example: 'категория'